GH-78: Fail CI on skipped tests (#80)
* Fail CI if skipped tests are present

Often during development work, we "skip" or "focus" certain tests while we are working on a particular feature.

And it's easy to forget and leave them in there, which causes not all tests to be run during CI.

We can adjust the lint config so that it yells at us if we forget and leave these in there.

In the case that we do have test cases that we _do_ want to skip, we can still add explicit enable / disable comments in order to allow this:

```js
/* eslint-disable jest/no-disabled-tests */
it.skip('the test we really do want to skip', () => {
  // ...
})
/* eslint-enable jest/no-disabled-tests */
```

---

Closes #78